Refactoring of common functions for middleware
